requirements.txt
algoliasearch==2.6.2
algoliasearch-django==2.0.0
asgiref==3.5.2
autopep8==1.6.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.13
djangorestframework==3.13.1
idna==3.3
pycodestyle==2.8.0
pytz==2022.1
requests==2.28.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.9
